requirements.txt
# Core Flask Application
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# HTTP Requests and API Integration
requests==2.31.0

# Configuration and Environment
PyYAML==6.0.1
python-dotenv==1.0.0

# Production Deployment
gunicorn==21.2.0 